Moved filename encode/decode functions into their own utilities package to remove code duplication. Other minor sonarqube issues.
