+++ CoreUtilities/src/au/gov/asd/tac/constellation/utilities/genericjsonio/JsonIO.java
@@ -17,6 +17,7 @@ package au.gov.asd.tac.constellation.utilities.genericjsonio;
 
 import au.gov.asd.tac.constellation.preferences.ApplicationPreferenceKeys;
 import au.gov.asd.tac.constellation.utilities.datastructure.Tuple;
+import au.gov.asd.tac.constellation.utilities.file.FilenameEncoder;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
@@ -134,7 +135,7 @@ public class JsonIO {
 
             // Create the file and write its contents. IF the file already exists, confirm from the user that they
             // wish to continue (and overwrite the existing file).
-            final File f = new File(prefDir, encode(fileName + FILE_EXT));
+            final File f = new File(prefDir, FilenameEncoder.encode(fileName + FILE_EXT));
             boolean go = true;
             if (f.exists()) {
                 final String msg = String.format("'%s' already exists. Do you want to overwrite it?", fileName);
@@ -214,7 +215,7 @@ public class JsonIO {
 
         // chop off ".json" from the filenames
         for (int i = 0; i < names.length; i++) {
-            final String nextName = decode(names[i].substring(0, names[i].length() - 5));
+            final String nextName = FilenameEncoder.decode(names[i].substring(0, names[i].length() - 5));
             if (nextName != null && !nextName.isEmpty()) {
                 names[i] = nextName;
                 // Hide any file prefix which the user didn't see when saving
@@ -232,7 +233,7 @@ public class JsonIO {
                 if (!filePrefix.isEmpty()) {
                     fileName = filePrefix.concat(fileName);
                 }
-                return mapper.readTree(new File(prefDir, encode(fileName) + FILE_EXT));
+                return mapper.readTree(new File(prefDir, FilenameEncoder.encode(fileName) + FILE_EXT));
             } catch (IOException ex) {
                 Exceptions.printStackTrace(ex);
             }
@@ -266,7 +267,7 @@ public class JsonIO {
         final File prefDir = new File(userDir, currentDir);
        
         if (filenameToDelete != null) {
-            final String encodedFilename = encode(currentPrefix.concat(filenameToDelete)) + FILE_EXT;
+            final String encodedFilename = FilenameEncoder.encode(currentPrefix.concat(filenameToDelete)) + FILE_EXT;
 
             // Loop through files in preference directory looking for one matching selected item
             // delete it when found
@@ -284,86 +285,5 @@ public class JsonIO {
                 }
             }
         }
-    }
-
-    /**
-     * Encode a String so it can be used as a filename. This method will
-     * essentially limit the characters that can be used in the filename, all
-     * other characters are escaped.
-     *
-     * @param s The String to be encoded.
-     *
-     * @return The encoded String.
-     */
-    public static String encode(final String s) {
-        final StringBuilder b = new StringBuilder();
-        for (final char c : s.toCharArray()) {
-            if (isValidFileCharacter(c)) {
-                b.append(c);
-            } else {
-                b.append(String.format("_%04x", (int) c));
-            }
-        }
-        return b.toString();
-    }
-
-    /**
-     * Decode a String that has been encoded by {@link encode(String)}.
-     *
-     * @param s The String to be decoded.
-     *
-     * @return The decoded String.
-     */
-    public static String decode(final String s) {
-        final StringBuilder b = new StringBuilder();
-        
-        // Loop through characters in source string. The normal case just adds
-        // the next character to the destination string, but if a '_ is found,
-        // it is treated asa marker which beigns a 4 character hex code - used
-        // by the encode method to represent characters identified as non
-        // printable.
-        int i = 0;
-        while (i < s.length()) {
-            final char c = s.charAt(i);
-            if (c != '_') {
-                b.append(c);
-                i++;
-            } else {
-                // An '_' was found, try to nibble off the next 4 characters
-                final String hex = s.substring(i + 1, Math.min(i + 5, s.length()));
-                if (hex.length() == 4) {
-                    // A full 4 characters could be nibbled off, which is expected
-                    // as all filenames generated by the encode method will add
-                    // this. So the 'bad' case would represent a filename added
-                    // to the config directory manually by a user.
-                    try {
-                        final int value = Integer.parseInt(hex, 16);
-                        b.append((char) value);
-                        i += 5; // Added 1 for the '_' and 4 for the digits
-                    } catch (final NumberFormatException ex) {
-                        // Hex string couldnt be converted to int
-                        return null;
-                    }
-                } else {
-                    // Hex string couldn't be extracted
-                    return null;
-                }
-            }
-        }
-        return b.toString();
-    }
-
-    /**
-     * *
-     * Returns true if the supplied character c is a valid character to use in a
-     * filename.
-     *
-     * @param c The character to test
-     * @return True if the character is able to be used in a filename.
-     */
-    public static boolean isValidFileCharacter(char c) {
-        return (c >= '0' && c <= '9') || (c >= 'A' && c <= 'Z')
-                || (c >= 'a' && c <= 'z') || c == ' '
-                || c == '-' || c == '.';
     }
 }
